Use Win32 synchronisation primitives and drop the global timer bump
Three changes to the threading and typing of the viewer core, plus the
comment and helper cleanup already sitting in the tree.

Font size

  ROISet carried its font size as long the whole way down, while
  IROIObject::GetFontSize already returned int32_t, so every renderer
  ended in a static_cast back. long is 32-bit on this target, so the
  values were never at risk — the type just said less than it could.
  The parameter is now int32_t from the public class through the layer
  to the four renderers, and the casts are gone. C++ callers must
  recompile: long and int mangle differently. The C ABI already took
  int32_t and is unaffected.

Atomics

  std::atomic is replaced by Interlocked* on volatile LONG / LONG64.
  Writes and read-modify-writes map directly; exchange(false) becomes
  InterlockedExchange, and the release/acquire pair guarding the status
  bar reservation is covered because Interlocked* is a full barrier.

  Reads stay plain volatile loads. The documented Interlocked idiom for
  a read is CompareExchange(&v, 0, 0), which is a locked read-modify-
  write that dirties the cache line the other thread is about to write
  — worse than what it replaces. MSVC defaults to /volatile:ms on x64,
  where a volatile read is an acquire and an aligned LONG read is
  atomic. The header records that dependency, and the x86 caveat on the
  64-bit member.

  The render FPS was a std::atomic<double>, which has no Interlocked
  equivalent short of punning the bit pattern. It now holds the frame
  interval in 100ns ticks instead — the unit SetWaitableTimer wants
  anyway, which also moves the division out of the frame loop.
  SetRenderFPS keeps its double signature.

Timer resolution

  timeBeginPeriod(1) was raised for the process lifetime, but nothing
  left depends on it. The frame timer asks for
  CREATE_WAITABLE_TIMER_HIGH_RESOLUTION precisely so it does not, the
  condition variable and the frame wait both block on INFINITE, and the
  engine clock is QueryPerformanceCounter. Measured at a 4.1667ms
  target: the high-resolution timer lands at 4.51ms without the bump
  and 4.35ms with it, while a plain waitable timer and Sleep both sit
  at 16.7ms without and 4.4ms with.

  So it is raised only when CreateFrameTimer fails to get the
  high-resolution timer and has to fall back, and released with the
  timer. The winmm dependency moves to the one file that uses it.

+++ D3D11ImageView/D3D11ImageView.h
@@ -141,10 +141,10 @@ public:
 	void WindowOverlayAdd(const Polyline2f* polylines, size_t count, const OverlayStyle& style);
 	void WindowOverlayAdd(const Polygon2f* polygons, size_t count, const OverlayStyle& style);
 
-	bool ROISet(const wchar_t* key, const wchar_t* name, const Rect2f& rect, COLORREF rgb, bool isMovable, bool isResizable, long fontSize);
-	bool ROISet(const wchar_t* key, const wchar_t* name, const Ellipse2f& ellipse, COLORREF rgb, bool isMovable, bool isResizable, long fontSize);
-	bool ROISet(const wchar_t* key, const wchar_t* name, const Circle2f& circle, COLORREF rgb, bool isMovable, bool isResizable, long fontSize);
-	bool ROISet(const wchar_t* key, const wchar_t* name, const Polygon2f& polygon, COLORREF rgb, bool isMovable, bool isResizable, long fontSize);
+	bool ROISet(const wchar_t* key, const wchar_t* name, const Rect2f& rect, COLORREF rgb, bool isMovable, bool isResizable, int32_t fontSize);
+	bool ROISet(const wchar_t* key, const wchar_t* name, const Ellipse2f& ellipse, COLORREF rgb, bool isMovable, bool isResizable, int32_t fontSize);
+	bool ROISet(const wchar_t* key, const wchar_t* name, const Circle2f& circle, COLORREF rgb, bool isMovable, bool isResizable, int32_t fontSize);
+	bool ROISet(const wchar_t* key, const wchar_t* name, const Polygon2f& polygon, COLORREF rgb, bool isMovable, bool isResizable, int32_t fontSize);
 	void ROIClear();
 
 	// 8bit 소스. bitDepth 를 받는 아래 오버로드에 8 을 넘기는 것과 같다.
